fix: content image
