+++ src/components/Footerv2/index.tsx
@@ -97,11 +97,6 @@ export const Footerv2: React.FunctionComponent<Footerv2 & React.HTMLAttributes<H
                   Practices Area
                 </span>
               </Link>
-              <Link href="/profile">
-                <span className={`text-white cursor-pointer ${router?.asPath?.includes('profile') && 'font-bold'}`}>
-                  Profile
-                </span>
-              </Link>
               <Link href="/ourattorneysmalvis">
                 <span
                   className={`text-white cursor-pointer ${router?.asPath?.includes('ourattorneysmalvis') && 'font-bold'}`}>
