+++ src/components/Modal/Default/index.tsx
@@ -117,14 +117,7 @@ export const DefaultModal: React.FunctionComponent<DefaultModal & React.HTMLAttr
                         Practices Area
                       </span>
                     </Link>
-                    <Link href="/profile">
-                      <span
-                        className={`text-white text-3xl cursor-pointer ${
-                          router?.asPath?.includes('profile') && 'font-bold'
-                        }`}>
-                        Profile
-                      </span>
-                    </Link>
+
                     <Link href="/ourattorneysmalvis">
                       <span
                         className={`text-white text-3xl cursor-pointer ${
