+++ src/pages/ourattorneysmalvis.tsx
@@ -1,4 +1,4 @@
-import { HeroMalvis, OurAttorneysMalvis } from '@/components/Chunks/OurAttorneysMalvis'
+import { HeroMalvis, ProfileMalvis, ProfileMalvis1, ProfileMalvis2 } from '@/components/Chunks/Profile'
 import { Footerv2 } from '@/components/Footerv2'
 import Head from 'next/head'
 
@@ -9,7 +9,9 @@ const OurAttorneyMalvis = () => {
         <title>MALVIS — Our Attorneys</title>
       </Head>
       <HeroMalvis open={false} />
-      <OurAttorneysMalvis />
+      <ProfileMalvis />
+      <ProfileMalvis1 />
+      <ProfileMalvis2 />
       <Footerv2 open={false} />
     </>
   )
